requirements.txt
requests
beautifulsoup4
python-dotenv
python-telegram-bot==20.0
selenium
webdriver-manager